pyproject.toml
[tool.poetry]
name = "hotel-project"
version = "0.1.0"
description = ""
authors = ["Мурад Камылов <kamylov@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-filter = "^23.2"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
